ads-github-fetch-all-upstreams (wip): extract total page count from HTTP "Link:" header
Building on the work that extracts data from the HTTP HEAD response to the
initial pagination request for /user/repos, we now examine the 'Link:' header
(if present) and extract the total number of pages from the 'page=N' url param
of the rel="last" entry.

Moves all of the previously inline code for parsing the HTTP response code (as
well as the new code for extracing the total pages from the 'Link:' header)
into a new function:

    f_get_gh_user_repos_pagination_count_or_die ()

CAVEAT: This version can only correctly extract data from the 'Link:' HTTP
header if it is on a single line. Empirical testing suggests this is the
always the case with the GitHub v3 API, but we should not assume it. It is
legit for HTTP header values to be represented across multiple lines, so we
should support that in the future. See RFC 2616 section 4.2 "Message Headers".