ads-github-fetch-all-upstreams (wip): use 'fork' flag to detect parent
The previous version of the program made separate calls to /repos/:owner/:repo
to look for the presence of the 'parent' object to determine whether a given
repository had a parent. That was being done becuase I had missed the
availability of this info in the output of /user/repos -- there it is
represented as the 'fork' flag.

This version now makes use of the 'fork' flag, and no longer makes any calls
to /repos/:owner/:repo