ads-github-fetch-all-upstreams (wip): add '--clone-if-missing' command line option
This option has the effect of cloning the user's GitHub repositories (using
the 'git-hub' tool) when a git working directory of the same name as the repo
is not found beneath the current working location.

Note that this feature works with the feature added earlier today of
specifying particular repos on the command line to limit the scope of the
action to just those specified (as opposed to cloning /all/ of the user's
missing repos).