ads-github-fetch-all-upstreams: add '-u' (--upstream-remote-if-missing) option
It may be the case that we have cloned one of our repositories manually, but
for whatever reason have not set up the 'upstream' remote yet; a "typical"
GitHub workflow would setup this remote to point to the parent GitHub
repository from which ours was cloned.

By default, we do not perform any action on any such working directories. But
if the '-u' ('--upstream-remote-if-missing') option is specified, we will
obtain the "clone url" of the parent repo and set it as the git remote named
"upstream".

I thought the 'git-hub' tool would just add the 'upstream' repo if invoked for
a repo for which an existing git working directory exists, but upon closer
reading I see that is not the case. For now we implement the functionality
inline, but I've left comments in the code around this because my current
thinking is that this capability might be a better fit as an enhancement to
'git-hub'.