ads-github-normalize-url: new command line tool
Normalize the URL provided on the command line. Emits normalized
representation of URL on stdout.

The URL emitted on stdout is sutiable for hashing or otherwise using as an
input to some function for generating an ID for the normalized URL.

The intent is to use this tool as from the upcoming 'ads-github-cache*'
tool(s).