ads-github-fetch-all-upstreams: avoid inadvertent joining of lines in REPO_DATA_LINES
There was a bug in our code that builds up the REPO_DATA_LINES list (across
successive pages of results from the GitHub /user/repo endpoint) that was
causing the first line of successive pages to be joined to the last line from
the previous page.

We now separate each page's results with an explicit newline character, and
later strip out empty lines to deal with the possibility of having introduced
spurious empty lines.

We also introduce some sanity checks in the code that iterates over the
REPO_DATA_LINES and parses out their individual fields.