ads-github-hash-url: new command line tool
Given a URL (or any value, really) either on the command line (or failing
that, on stdin) computes an ID for the URL.

Since there are many textual representation for the resource indicated by a
particular URL, a typical caller would "normalize" the URL (using
'ads-github-normalize-url', or similar) before requesting that an ID be made
out of it.

The ID is currently computed using the SHA-3 512 bit algorithm variant.