add file: doc/ads-github-tools-design-notes.org
The ads-github-tools are already usable and proving themselves to be valuable
time savers, but they are only in their infancy (there is still no use of
caching, for instance), are only documented in the source code, and each needs
to be manually installed. I'm capturing random tidbits of where I'm taking the
tools in this file.