ads-github-fetch-all-upstreams: move first-pass cli opts parsing into function
This changeset builds upon the previous by moving the first-pass command line
options rewriting activity into its own function:

    f_rewrite_command_line_or_die ()

This is a step in the right direction, but the above function is still
hard-coded for this particular program and relies on the caller having
established an array variable named 'MY_COMMAND_LINE_OPTIONS' with the
program's command line options.

A few new helper functions are also introduced:

    f_is_array ( MAYBE_ARRAY_NAME ) - Indicates (via its return value) whether
    or not the provided MAYBE_ARRAY_NAME parameter is the name of a bash
    array.

    f_shift ( ARRAY_NAME [, N] ) - Works similarly to bash's built-in 'shift'
    command, but operates on the named array rather than $@. If N is not
    given, it is assumed to be 1. Returns success unless N is negative or
    greater than the number of elements in the named array.

    f_maybe_increase_verbosity () - encapsulates previously inline processing
    into a function to allow it to be invoked from the new command line
    rewriting preprocessing as well as from the original location. Also
    generalizes the code a bit.

Note, too, that we're now using a consistent trace-level output message
format.