fix #1: ads-github-fetch-all-upstreams: complain if specified repo not processed
This changeset fixes 'ads-github-tools' github issue #1. If a repository is
explicitly named on the `'ads-github-fetch-all-upstreams'` command line and is
not actually processed, then the program emits an error message and exits with
a non-zero status.

This helps avoid the false-warm-and-fuzzy from silent non-action when the user
fat-fingers one or more repo names on the command line.