fix #2: ads-github-merge-all-upstreams: introduce '-L' (--stop-upon-local-mods) opt
With this change the default behavior of the program is altered to
automatically skip over any repo that is found to have any uncommitted local
mods. A warning message is printed (indicating that the repo is being skipped
over), but the program will carry on and later exit with a success status of
zero if no other problems are encountered.

The new `'-L'` (`--stop-upon-local-mods`) command line option has been added
to allow the user to explicitly request that the program stop processing upon
encountering a repo with uncommitted local mods. If the program finds any,
then the exit status will be `99` to allow the invoking parent process to
detect the scenario, as well.

This changeset also tightens-up our exit status detection code when obtaining
the remote branch names. We eliminate the intermittent failures the author was
seeing by /avoiding/ treating as an error a `SIGPIPE` received by either our
`'git branch --remotes'` command or the down-pipe `'sed'` command.