fix #3: ads-github-merge-all-upstreams: introduce '-k' (--keep-going) opt
If we get an error while attempting to perform a fast-forward merge on a given
repo, the default behavior of the program is (and always has been) to exit
with an error message (and with a non-zero error status, of course). If you
are not expecting the problem with the fast-forward merge, then that is
exactly the behavior that you want.

If you /are/ expecting it, though, you will be annoyed that the program has
stopped instead of just moving onto the next repo. Enter the `'-k'` option; if
specified, the program will stop processing the repo in which it could not
perform the fast-forward merge, but will go on to the next repo.

If such a condition is encountered, the program will still exit with an error
status; it will just do so later. The very last message printed by the program
will then be something like:

    ads-github-merge-all-upstreams (error): one or more errors encountered with '-k' (--keep-going) specified (see error output above for details); bailing out

This changeset also introduces a bit more structure around error codes. This
is mainly an internal processing detail between the parent process and
subshells, but there is a user-visible change in that the `'-L'`
(`--stop-upon-local-mods`) option will cause the program to exit with a status
of `100` rather than `99` if it encounters its condition. Since I'm the only
user AFAIK I'm not worried about breaking anything with this change. See
comments at the top of the source file for details.