more changes for #5: add support for generating perl-based progs
Currently the only perl-based prog is 'ads-github-normalize-url'; that program
is now generated from the template of the same name that has a '.in'
extension.

The build also generates a man page for the prog from perldoc embedded at the
end of the file. Any future perl-based prorgrams would follow the same model
and plug into the build in the same way. The man page content was just scraped
together quickly so it would be at least marginally useful.

The current build produces programs that have the project version number
suffixed to them. This is useful for having multiple different versions
installed on a given system side-by-side while retaining the ability to invoke
specific versions. It's not too friendly, though, when the desired use case is
"invoked the latest version" or "invoke the latest stable version". Expect
this to evolve a bit...