more changes for #5: bash scripts now part of autoconfiscated build
The autotools-based build now works (or gives the appearance of working). I'm
still cleaning things up and tinkering, but the core of the functionality is
there and the project builds fine.

I'm still working on deciding how I want to handle progs with verison numbers
append to their program names. I'm leaning in the direction that developers
who want to have multiple versions of the tools installed side-by-side will be
savvy enough to use different `--prefix=...` options at configure-time. Either
way, there's still some messiness in the Makefile.am files related to this.

Also added support for the various PACKAGE_FOO output variables and
preprocessort symbols emitted by AC_INIT. Previously we supported only PACKAGE
and VERSION, and did anything else manually.

The Perl and Bash programs both use POD ("plain old documentation") formatted
text for their respective documentation, and the build uses perldoc(1) to
generate the manpages. The build will bomb out if there is no documentation
for a given tool (this is a feature), so this changeset adds some minimal man
pages for all remaining tools that did not have them previously.

The `ads-github-tools` programs have relatively long names. When generating
the man pages, without intervention the default behavior would be for fields
in the top-of-page banner run together and become unreadable.

The general format is:

```
    <TOOL_NAME>(N)                 ads-github-tools                 <TOOL_NAME>(N)
```
However, with some of our long tool names, we would get something like this on
an 80-column terminal (or when `$MANWIDTH` is set to `80`, which is common
practice):
```
    ads-github-fetch-all-upstreams(ads-github-tooads-github-fetch-all-upstreams(1)
```
We work around this by slicing off the following prefixes from tool names as
printed in the top-of-page banner in the man pages. This changes the above
output to look like:
```
    fetch-all-upstreams(1)         ads-github-tools         fetch-all-upstreams(1)
```
The author is not thrilled about this; suggestions for alternative approaches
are welcome.