more changes for #5: check-in (most) autotools-generated files
There are generally two schools of thought on whether generated files should
be checked into source control, and there are tradeoffs involved with both
practices.

The author of the ads-github-tools (yours truly) does not want to receive
patches to any of the autotools-generated files, but other than that finds
value in having the generated files in the git repo.

Having the generated files in the repo allows for building the project on a
host that does not have the full autotools toolchain, which is useful in some
situations. Having the generated files in the repo also allows differences in
those files to be compared over time, which is useful for evolving a project.

The author typically commits autotools-generated files in a standalone
changeset for the purpose to help keep changeset diffs readable.