more changes for #5: autotools-based build is now fully functional
There are still some things that need cleaning up, but the autotools-based
build is now fully functional. That is, it is complete enough to generate the
programs in the project, generate their man pages, and produce a source code
distribution tarball via 'make distcheck'.

There are some known deficiencies; these will be addressed over time. The most
severe is probably that various external tools (`git`, `git-hub`, `jq`,
others) are not yet checked for at configure time.