fixes #10: all ads-github-tools progs now use (by default) tools found at configure time
This changeset simply contains all of the regenerated autotools-generated files.

All external tools are now searched for at configure time. If not found the
configure process bails out with an error.

The found tools are filtered into the generated programs at build time, but in
such a way that users are able to override individual tools by specifying the
path to a specific tool in an environment variable named for the tool (with
hyphen chars changed to underscore chars); this feature is not yet documented
in any of the man pages because it is expected to be a developer-only feature.