fixes #12 (*-fetch-all-upstreams, *-merge-all-upstreams): ignore org-owned repos
I learned over the weekend that the data returned from `/user/repos/` includes
not only the repositories for which the authenticated user is an owner and his
personal forks of other repos, but also all organization-owned repos that the
user has access to (regardless of whether or not he has a personal fork of
them).

This means that a fundamental assumption that the
`'ads-github-fetch-all-upstreams'` and `'ads-github-merge-all-upstreams'`
programs were making about the repos list obtained from that endpoint --
namely that it represented repos that the user either owned or had forked from
others -- is incorrect. The data returned is (potentially) a superset of that
data.

Users who are not part of any organizations would not be affected, as the
author was not prior to becoming a member of an organization and tripping over
the bug.

The buggy behavior of the `'ads-github-fetch-all-upstreams'` tool was to treat
all data from `/user/repos/` as personal forks. The organization-owned repos
that are listed as being available are simply a different beast. The
`'ads-github-fetch-all-upstreams'` tool, in assuming that all repos in the
obtained list, attempted to clone the other org repos that the user had
available to him for forking, but for which he had no personal fork actually
created. The program then exited with an error.

This changeset updates the `'ads-github-fetch-all-upstreams'` and
`'ads-github-merge-all-upstreams'` progs to work only with repos with owner
type `"User"` (we are ignoring those with owner type `"Organization"`). This
works for the author's use case, but may be more limiting than we want to be
in the long-term. The approach works because a user's personal fork of an
organization-owned repo shows up in the `/user/repos/` list in the way that
the program was expecting previously. This means that some repos can show up
twice in the list:

   1. once with the organization indicated as the owner (this is the repo
      being listed as available to the user by virtue of his membership in
      the organization)

   2. once with the user indicated as the owner (this is the user's personal
      fork of the repo).

These changes have the (probably) undesirable effect of preventing those
programs from working directly on organization-owned repos. I author need to
further review how those are intended to work with cloning/merging/pushing,
but if this turns out to be an artificial limitation I'll be interested in
having it fixed.

Patches/pull requests welcome!