[1 of 2] fixes #13: bash-based progs now have simplified (but full featured) impl
Prior to this changeset, the bash-based progs in `ads-github-tools` provided
short- and long-form options processing (with bundling for short-form opts) by
using the builtin `getopts` operator for the short-form bundling and then
converting those to the equivalent long-form opts (rewriting the command line
in the process). Only once all of that was done did we get down to the
business of actually parsing the long-form opts. Not only was the
implementation inefficient, but there was a ton of code involved, some of it a
bit sketchy.

This changeset simplifies the implementation by inverting that processing so
that long-form options are special-cased instead. We still use bash's builtin
`getopts` operator, but now we recognize long-form options as we encounter
them and just fix up the normal variables involved with `getopts` processing
(`$OPTARG`, `$OPTIND`, and our own `$opt`) to accommodate them. Not only is
this more efficient, it's also much less code while still being "full
featured" (we still get bundling of short-form options, and long-form options
that take an argument can be specified in both of these forms):
```
    --foo=FOOVAL
    --foo FOOVAL
```

The following are the programs from `ads-github-tools` that are implemented in
bash; all have been updated as describe above:

   * `ads-github-fetch-all-upstreams`
   * `ads-github-hash-url`
   * `ads-github-merge-all-upstreams`
   * `ads-github-show-rate-limits`

Note that our updated approach makes use of associative arrays, which first
appeared in `bash-4.0`, so we have also updated our `'configure'` process to
ensure that bash 4.x or newer is being used. Since `bash-4.0` was released way
back in `2009-02` the author is hoping that this will not create a hardship
for anybody.

Finally, the removed code contains two bash functions that are of general
utility that we may resurrect in some other context in the future. We're
mentioning them in this commit message to make them easier to find should that
happen:

   * `f_is_array(...)` -- "Indicates whether or not the provided
     MAYBE_ARRAY_NAME parameter is the name of a bash array."

   * `f_shift(...)` -- "Works similarly to bash's built-in 'shift' command,
     but operates on the named array rather than $@.

     Usage: f_shift ARRAY_NAME [n]

     If N is not given, it is assumed to be 1.

     Returns success unless N is negative or greater than the number of
     elements in the named array."