fixes #14: remove pre-autotools hard-coded cruft
Prior to this changeset the various programs in the `ads-github-tools` project
ha some commented-out, hard-coded cruft from the days before the project had a
proper build based on the GNU autotools. For example, we saw lots of this:
```
    # FIXME: one day this will be filtered in at build time
    # declare -r MAINTAINER='Alan D. Salewski <salewski@att.net>'
    # declare -r MAINTAINER='@DFLT_MAINTAINER_FULL@'
    declare -r MAINTAINER='@PACKAGE_BUGREPORT@'

    # FIXME: one day this will be filtered in at build time
    # declare -r VERSION='0.0.1'
    declare -r VERSION='@VERSION@'
```
The comments, obviously, were anticipating that the values for those variables
would be filtered-in at build time. Now that the build has been established,
however, we have gone through cleaned out the crufty commented-out bits.