fixes #11: introduce "aggressive push" mode control ('-pp') in ads-github-merge-all-upstreams
This changeset alters the behavior of the existing '-p' (--push) option of the
`'ads-github-merge-all-upstreams'` program.

The old behavior, when one or more `'-p'` options were specified, was to
always perform a `'git push ...'` operation. This was very simple to
understand, but also came with a fair amount of inefficiency. When working
with a large number of repositories, most of which did not have any newly
merged changes that needed to be pushed, the push operation was performed
anyway (with all of the network overhead that implies).

The new behavior is to support two different push modes: "basic" and "aggressive".

The "aggressive" mode is identical to the old behavior, but the user now needs
to specify two (or more) `'-p'` options to request the (potentially
inefficient) behavior explicitly.

With only one `'-p'` option, the new behavior is to only perform a `'git push ...'`
operation if changes were actually merged into a given repo during the current
invocation of the program. See `ads-github-merge-all-upstreams(1)` for
additional details and motivating program usage patterns.