fixes #16 (ads-github-merge-all-upstreams): qualify push refspec as branch
If a repository had a tag name that matched the default branch name (typically
"master"), then the previous version of the program would not be able to push
the merged changes to the origin repository because the git refspec used
internally was ambiguous. The program has been enhanced to qualify the default
branch name as "heads/<BRANCH>" rather than just "<BRANCH>" to remove this
ambiguity.