fixes #21 (ads-github-merge-all-upstreams): no longer chokes when it encounters a "detached HEAD"
Prior to this changeset the program choked (when the repo directory was
initially in a "detached HEAD" state) upon attempting to restore the working
directory to the state it was in prior to the program switching to the default
branch. This happened because the code only knew how to deal with a working
directory "attached" to a named branch. Since the repo was originally in the
detached HEAD state, the program did not have a named branch to which it could
switch back.

Rather than attempt to switch back to an explicitly named branch, the code now
basically does this:
```
    git checkout @{-1}
```
which you can read as: "where HEAD used to be one move ago (before we
checked-out the default branch)".

The impact of the original problem was that the
`'ads-github-merge-all-upstreams'` program stopped dead in its tracks and left
the repo directory with the default branch checked out.

The workaround was to simply re-run the program leaving that repo with the
default branch checked out. Because the program then found the repo with the
default branch already checked out, it did not attempt to swicth the branch,
so also did not need to attempt to restore it to its original state. Once the
program completed, the user could then manually check out the tag (or
whatever) to then put the repo back into the original detached HEAD state (if
so desired).

If something goes wrong restoring the working directory to the detached HEAD
state, then any commits the user has made to the detached HEAD that are not
referenced from some branch (or, eventually, the reflog) will be "unreachable
objects" and in danger of (eventually) being cleaned up by git's periodic
garbage collection. See the documentation for `'gc.reflogExpireUnreachable'`
and friends in `git-gc(1)` for more on configuring that behavior. This program
makes an effort to detect situations that might lead to this situation prior
to changing to the default branch in an attempt to avoid such a
scenario. Nevertheless, if the program is not able to restore the detached
HEAD state, then the repo will end up on the default branch. This is treated
as an error and the program stops, but it is up to the user to do whatever is
necessary to prevent garbage collection of commits from the detached HEAD. In
most circumstances, the user will want to go back to the detached HEAD by
simply invoking the `'git checkout @{-1}'` command noted above from within the
repo's working directory.