[1 of 2] fixes #23 (ads-github-fetch-all-upstreams): note all unprocessed repo names in err msg
Prior to this changeset, when a user fat-fingered more than one repo name on
the command line, the program printed an error message that named only the
first unprocessed repo name that was detected.

The current changeset improves on that behavior to name /all/ unprocessed repo
names. This gives a better indication to the user about the size of the error,
and also gives the user the opportunity to correct all such errors at once
(rather than playing whack-a-mole fixing one at a time and having successive
runs of the program fail on the remaining repos that were not named in the
error message).