ads-github-tools-0.2.0 release prep
- updated version in 'configure.ac'
- set the release date in 'NEWS'
- referenced version 0.2.0 as "latest", etc., in README.md
- regenerated 'configure'

Note that the hyperlinks that reference version `0.2.0`-specific locations
will not be valid until release artifacts are uploaded to their referenced
locations (which will happen around the same time that this changeset is
pushed up to GitHub).

This changeset also updates the `README.md` file with blurbs highlighting the
new `--missing-only` option of `'ads-github-fetch-all-upstreams'`, as it is
expected to be used in one of the three most common invocations.