fixes #25: ads-github-show-rate-limits(1) now correctly documents data retrieval mechanism
The ads-github-show-rate-limits(1) man page previously mis-documented the
underlying mechanism used by the tool to obtain the user's GitHub API rate
limits information. It was documented to obtain that information from the HTTP
response headers rather than the JSON response body from the '/rate_limit'
service endpoint.

While per-endpoint calls do allow for the possibility of extracting GitHub
service limit information out of the HTTP response headers (e.g., see the
["Rate Limiting"][docs_rate_limiting] section of the GitHub API
documentation), the 'ads-github-tools-show-rate-limits' program is actually
making a call to the GitHub [/rate_limit][endpoint_rate_limit] API endpoint
and extracting the data that it displays out of the JSON returned in the HTTP
response body.

This changeset corrects the record to accurately reflect the underlying
mechanism that is used to retrieve the user's current GitHub service limits
information.

[docs_rate_limiting]:  https://developer.github.com/v3/#rate-limiting
[endpoint_rate_limit]: https://developer.github.com/v3/rate_limit/