[1 of 2] fixes #6: ads-github-show-rate-limits: add '-u' (--utc or --universal) option
This changeset enhances the `'ads-github-show-rate-limits'` program with a new
`'-u'` (`--utc` or `--universal`) option that allows the user to request that
time fields be output in human-readable [(RFC 3339)][] formatted UTC (as
opposed to the default seconds since the Unix epoch, or (with the `'-h'`
(`--human-readable`) option) human-readable local time.

See `ads-github-show-rate-limits(1)` for details.

Note that this changeset is also the first in the `ads-github-tools`
`0.2.1-snapshot` line of development to introduce a new feature, so
(consistent with [semver 2.0][]) we bump up the MINOR field so that our new
in-development version is `0.3.0-snapshot`.

[RFC 3339]:   https://www.ietf.org/rfc/rfc3339.txt
[semver 2.0]: http://semver.org/