fixes #8: ads-github-show-rate-limits: add '-O' (--output-format=WORD) option
The `'ads-github-show-rate-limits'` program has a new `'-O'` (`--output-format=WORD`)
option that allows the user to request different output formats. Currently the
only two output formats accepted are `text` (the default) and `json`.

The default output of the program has not changed. It is still
whitespace-separated columns of text. Specifying the new
`--output-format=text` option is just an explicit way of requesting the
default behavior.

See `ads-github-show-rate-limits(1)` for full details.