fixes #45 [2 of 2]: automake-1.16 migration: regenerate auto-generated files
This changeset simply commits the autotools-generated files as part of our
migration from automake-1.14 to automake-1.16.

The command used to cause the regeneration was:
```
    $ ./bootstrap --force --install
```

Note that this update not only regenerates `configure` and the various
`Makefile.in` files, but also pulls in updated versions of
`build-aux/install-sh` and `build-aux/missing`.