teach contrib/ads-gh-repo to delete repos
If the credentials allow it, and the DO_FORCE_REPLACE flag is set, then the
program will use the GitHub API to delete the specified repository.

Note that DO_FORCE_REPLACE is currently enabled -- hard-coded. In the future
it will be controlled via a command line option.

Note, too, that the permissions associated with most sane access credentials
will not allow remote processes to delete repositories, so this functionality
is not likely to be used often (but who knows...).

IMHO, I think the risk of losing precious data is too high; for my purposes,
I'll stick with forcing the user to manually delete repos through the GitHub
web UI.