fixes #42: adapt for github deprecating password authn, OAuth authn & authz APIs
No behavioral changes needed. Updated docs in various places to
inform users of the GitHub 2020-11-13 hard-cutoff date, and to
strongly urge the use of personal access tokens over a
password. This is probably just a formality; I'm not aware of
anybody actually using a password rather than a personal access
token.