fixes #44 [1 of 2]: ads-github-show-rate-limits: http response code not checked
The 'ads-github-show-rate-limits' program is now careful to check
not only the curl(1) exit status of the 'GET /rate_limit' request,
but also the HTTP response status.

Previously it was obtaining the JSON response on the curl(1) stdout,
and had no way to determine what the HTTP response code was.

The new behavior is to follow our pattern of saving the raw response
output to a temporary file, and having curl write the HTTP response
code to stdout; this allows us to deal with both aspects
independently.