fixes #33 [1 of 2]: fetch-all-upstreams: rework all curl, curl->jq interactions
The 'ads-github-fetch-all-upstreams' tool was previously using the
default behavior of curl(1) to invoke GitHub v3 API endpoints and
emit the retrieved responses on stdout, feeding that output in Unix
pipelines to other programs (mostly to jq(1)).

While that approach was fast and convenient when originally
implemented, it also had the downside of preventing the program from
properly detecting all failures because it could not check the HTTP
response codes.

We now follow our pattern (used elsehwere in the ads-github-tools)
of configuring curl to write the response body to a temporary file
named by a well-known global variable, and emit the HTTP response
code on stdout. The calling code is thus able to check the curl exit
status, the emitted HTTP response code, and the size and content of
the response body independently of one another. This makes the
program more robust in general, and avoids complaints emitting from
the downstream jq(1) process that should have been handled upstream
of it.

These functions are newly introduced:

    f_file_ends_with_newline()
    f_initialize_curl_default_opts()
    f_build_repo_data_lines_or_die()

This existing function has been modified to conform with the above
behavioral description:

    f_get_gh_user_repos_pagination_count_or_die()